Ajout de DTOS et leur gestion dans les services

+++ src/main/java/io/ipme/neptunes/Controller/ThemeController.java
@@ -1,6 +1,5 @@
 package io.ipme.neptunes.Controller;
 
-import io.ipme.neptunes.Model.Theme;
 import io.ipme.neptunes.Service.ThemeService;
 import io.ipme.neptunes.Service.dto.ThemeDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
+import java.util.List;
 
 @RestController
 @RequestMapping("themes")
@@ -22,7 +21,7 @@ public class ThemeController {
     }
 
     @GetMapping
-    public ResponseEntity<ArrayList<ThemeDTO>> findAll() {
+    public ResponseEntity<List<ThemeDTO>> findAll() {
         try {
             return ResponseEntity.ok(themeService.findAll());
         } catch (Exception e) {
@@ -31,28 +30,52 @@ public class ThemeController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<Theme> findById(@PathVariable Integer id) {
+    public ResponseEntity<ThemeDTO> findById(@PathVariable Integer id) {
         try {
-            return ResponseEntity.ok(themeService.findById(id));
+            if (id != null) {
+                return ResponseEntity.ok(themeService.findById(id));
+            }
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 
     @PostMapping
-    public ResponseEntity createTheme(@RequestBody Theme theme) {
-        themeService.createTheme(theme);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<?> createTheme(@RequestBody ThemeDTO themeDTO) {
+        try {
+            if (themeDTO != null) {
+                return ResponseEntity.ok(themeService.createTheme(themeDTO));
+            }
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
     }
 
     @DeleteMapping("{id}")
-    public void deleteTheme(@PathVariable Integer id) {
-        themeService.deleteTheme(id);
+    public ResponseEntity<String> deleteTheme(@PathVariable Integer id) {
+        try {
+            if (id != null) {
+                themeService.deleteTheme(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
     }
 
     @PatchMapping("{id}")
-    public void updateTheme(@RequestBody Theme theme, @PathVariable Integer id) {
-        themeService.updateTheme(theme, id);
+    public ResponseEntity<?> updateTheme(@PathVariable Integer id, @RequestBody ThemeDTO themeDTO) {
+        try {
+            if (id != null && themeDTO != null) {
+                return ResponseEntity.ok(themeService.updateTheme(id, themeDTO));
+            }
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
     }
 
 }
